refactoring de metodo Handler saveMovie OK

+++ src/main/java/app/disney/service/implement/MovieServiceImplemet.java
@@ -11,9 +11,13 @@ import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+import app.disney.DTO.MovieDTO;
 import app.disney.DTO.SearchMovieDTO;
 import app.disney.entitys.Movie;
+import app.disney.exceptions.ExistingNameException;
+import app.disney.exceptions.NotFoundException;
 import app.disney.repository.IMovieRepository;
+import app.disney.service.IGenderService;
 import app.disney.service.IMovieService;
 import app.disney.specification.MovieSpecification;
 import app.disney.util.IMapper;
@@ -29,6 +33,9 @@ public class MovieServiceImplemet implements IMovieService {
 
 	@Autowired
 	private MovieSpecification spec;
+	
+	@Autowired
+	private IGenderService genderService;
 
 
 
@@ -39,18 +46,49 @@ public class MovieServiceImplemet implements IMovieService {
 	}
 
 	@Override
-	public Movie saveMovie(Movie movie) {
-		return movieRepo.save(movie);
+	public void  saveMovie(MovieDTO newMovie, 
+			               MultipartFile imagen, 
+			               String gender) {
+		Movie movie= mapping.mappingMovieDTOToEntity(newMovie);
+		validateName(movie.getTitle());
+		validateImagenAndListMovie(movie, imagen, gender);
+		
+		System.out.println("////////////////////////////////////////");
+		movieRepo.save(movie);
 	}
-
+	
+	
+	
+	
+	@Override
+	public void validateName(String title) {
+		if (movieRepo.findByTitleIgnoreCase(title) != null) 
+			throw new ExistingNameException("TITLE  : " + title);
+	}
+	
+	@Override
+	public void validateImagenAndListMovie(Movie movie, 
+			                               MultipartFile imagen, 
+			                               String gender) {
+		
+		if (imagen != null) saveImg(movie,imagen);
+		if (gender != null) movie.setGender(genderService.getByNameGender(gender));
+		
+		
+	}
+	
 	@Override
 	public Movie getMovieById(Integer id) {
-		return movieRepo.findById(id).get();
+		return movieRepo.findById(id).orElseThrow(() -> new NotFoundException("ID : " + id));
 	}
 
 	@Override
 	public void deleteMovieById(Integer id) {
-		movieRepo.deleteById(id);
+		
+		if (movieRepo.findById(id).isPresent())
+			movieRepo.deleteById(id);
+		else
+			throw new NotFoundException("ID : " + id);
 	}
 
 	/* BUSQUEDAS */
@@ -69,7 +107,7 @@ public class MovieServiceImplemet implements IMovieService {
 
 	///////////////////////////////////////////////////////////////
 	@Override
-	public void saveImg(MultipartFile imagen) {
+	public void saveImg(Movie movie, MultipartFile imagen) {
 		Path directorioImagenes = Paths.get("src//main//resources//static/imgMovies");
 		String rutaAbsoluta = directorioImagenes.toFile().getAbsolutePath();
 
@@ -77,13 +115,14 @@ public class MovieServiceImplemet implements IMovieService {
 			byte[] bytesImg = imagen.getBytes();
 			Path rutaCompleta = Paths.get(rutaAbsoluta + "//" + imagen.getOriginalFilename());
 			Files.write(rutaCompleta, bytesImg);
+			// SET Imagen
+			movie.setImgMovie(imagen.getOriginalFilename());
+			
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
-		
 	}
 	
-
 	@Override
 	public List<Movie> getAllMovieBySpec(Specification<Movie> spec) {
           return movieRepo.findAll(spec);
@@ -110,5 +149,8 @@ public class MovieServiceImplemet implements IMovieService {
 
 		}
 	}
+
+
+
 
 }
